pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "terraform-plan-reviewer"
version = "0.1.0"
description = "A CLI tool to summarize Terraform plan JSON files using GPT"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
dependencies = [
    "click",
    "openai",
    "python-dotenv",
    "rich"
]

[project.scripts]
tf-reviewer = "terraform_plan_reviewer.cli:analyze"
